Inform modules when packets are dropped, and try to make sure that the core doesn't panic when such unpleasant events occur.
At the moment, only the sk98 sniffer reports drops correctly, but
modules should be informed of drops in the core when using any
sniffer.

Also, none of the modules do anything useful with the drop information
yet.

Note that this commit changes the prototype for the next method of
modules.  I've converted all of the modules in the tree over, and it
should be a trivial fix for any out of tree ones.