customize is finished for now.  Need to know the main window modules and how to use them.
dashboard is currently being fixed.  urlargs are different for each module
and it may be smart to have these exported by como.

node.class:  added some new function to return the modules in the conf file

live.css : added a vert align on the rightcontent to force everything to the
top