comolive.conf removed line containing the NOSHOW because it is now configurable from customize.php
customize.php
removed the adding of a file that doesn't exist into the node.class

changes to the rest of the files make the customize work with the
selecting of right hand side elements (alert, topports, etc) and the
modules that can be selected on the mainstage (traffic, application, etc.).
Now the module in como must export this information to appear on the live
site.