Major change in the memory managenent. It is now entirely done in shared memory and modules can allocate memory in the shared space so that all callbacks have access to the same memory region.
New functions to be used in the modules include: mdl_mem_alloc, mdl_mem_free
that allocate memory in shared/private space depending on the specific
callback where they have called. In the init() callback all memory allocations
are persistent, i.e. they survive the periodic memory flush done by CAPTURE
and they are visible to all processes when the module is loaded.

All callbacks have now changes and they now have a first argument
"void * self" that can be used to access the private state of the
module. There is no more any need to declare static variables in the
module code.

Note that right now the mechanism to specify indesc and outdesc is
broken and will be fixed as soon as possible.