Port-scanner Detection Module.
Author: Derek Hossack Lopez (UPC, Barcelona) 

This module detects port scanners by tracking the TCP 3-way handshake at 
the beginning of a TCP connection.

The output of this module is a packet-level trace of the hosts
classified as port scanners, which can be directly pipelined to the 
flow-reassembly module.

This module is inspired on the method described in "Autograph: Toward 
Automated, Distributed Worm Signature Detection" by Hyang-Ak Kim and 
Brad Karp.

Configuration example (see source code for extra args):

module "scanner-detector"
    description "Port-scanner detector"
    hashsize 100000
end