Worm Signature Generation Module.
Author: Derek Hossack Lopez (UPC, Barcelona)

This module automatically generates worm signatures from the traffic it
receives using the method described in "Autograph: Toward Automated,
Distributed Worm Signature Detection" by Hyang-Ak Kim and Brad Karp.

This module is supposed to use as input the output from 
'scanner-detector' and 'flow-reassembly' modules, so that the traffic 
under analysis consists of reassembled TCP flows coming from suspicious 
IP addresses.

The flows are stored in the export hash table grouped by destination 
port. Periodically, the module does content-based partitioning of the 
suspicious flows with the same destination port, and measures the 
prevalence of each content block. The most prevalent blocks are then 
chosen as signatures, which can be obtained in Snort format by querying 
the module.

This module requires libssl. If libssl is not present, the module is not
compiled.